move RegistrationPromo to its own file
